Put ExtraTools and ExtraOptions in distinct model file
