+++ projects/ngx-testing-extra/src/lib/components/extra-bed/create-extra-bed.ts
@@ -2,8 +2,9 @@ import { Type } from '@angular/core';
 import { ComponentFixture } from '@angular/core/testing';
 import { buildActionTools } from './action-tools';
 import { ExtraBedFactory } from './extra-bed-factory';
+import { ExtraOptions } from './models';
 import { ActionTools } from './models/action-tools.model';
-import { ExtraBed, ExtraCb, ExtraOptions } from './models/extra-bed.models';
+import { ExtraBed, ExtraCb } from './models/extra-bed.models';
 import { QueryTools } from './models/query-tools.model';
 import { buildQueryTools } from './query-tools';
