+++ projects/ngx-testing-extra/src/lib/components/extra-bed/models/extra-options.model.ts
@@ -1,0 +1,6 @@
+export interface ExtraOptions {
+  /**
+   * @default true
+   */
+  startDetectChanges?: boolean;
+}
